+++ doc/api/template.py
@@ -35,6 +35,7 @@ MODULES = [
     'http.request',
     'http.robots',
     'http.stream',
+    'http.proxy',
     'http.util',
     'http.web',
     'item',