+++ doc/changelog.rst
@@ -2,6 +2,7 @@
 What's New
 ==========
 
+* Support for HTTP/HTTPS proxies but no HTTPS tunnelling support. Wpull will refuse to start without the insecure override option. Note that if authentication and WARC file is enabled, the username and password is recorded into the WARC file.
 * API:
 
     ``document`` and ``scraper`` were put into their own packages.